Organize file structure and start fetch request for book search
